+++ src/Symfony/Component/Mailer/Bridge/Postmark/Tests/Transport/PostmarkApiTransportTest.php
@@ -21,10 +21,12 @@ use Symfony\Component\Mailer\Bridge\Postmark\Transport\MessageStreamHeader;
 use Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkApiTransport;
 use Symfony\Component\Mailer\Envelope;
 use Symfony\Component\Mailer\Exception\HttpTransportException;
+use Symfony\Component\Mailer\Exception\InvalidArgumentException;
 use Symfony\Component\Mailer\Exception\TransportException;
 use Symfony\Component\Mailer\Header\MetadataHeader;
 use Symfony\Component\Mailer\Header\TagHeader;
 use Symfony\Component\Mailer\Header\TrackingHeader;
+use Symfony\Component\Mailer\RemoteTemplateEmail;
 use Symfony\Component\Mime\Address;
 use Symfony\Component\Mime\Email;
 use Symfony\Contracts\HttpClient\ResponseInterface;
@@ -112,6 +114,68 @@ class PostmarkApiTransportTest extends TestCase
         $message = $transport->send($mail);
 
         $this->assertSame('foobar', $message->getMessageId());
+    }
+
+    public function testSendRemoteTemplate()
+    {
+        $client = new MockHttpClient(function (string $method, string $url, array $options): ResponseInterface {
+            $this->assertSame('POST', $method);
+            $this->assertSame('https://api.postmarkapp.com/email/withTemplate', $url);
+
+            $body = json_decode($options['body'], true);
+            $this->assertSame(12345, $body['TemplateId']);
+            $this->assertSame(['name' => 'Fabien'], $body['TemplateModel']);
+            $this->assertArrayNotHasKey('TemplateAlias', $body);
+            $this->assertArrayNotHasKey('Subject', $body);
+            $this->assertArrayNotHasKey('TextBody', $body);
+            $this->assertArrayNotHasKey('HtmlBody', $body);
+
+            return new JsonMockResponse(['MessageID' => 'foobar'], [
+                'http_code' => 200,
+            ]);
+        });
+
+        $transport = new PostmarkApiTransport('KEY', $client);
+
+        $mail = (new RemoteTemplateEmail())
+            ->to(new Address('saif.gmati@symfony.com', 'Saif Eddin'))
+            ->from(new Address('fabpot@symfony.com', 'Fabien'))
+            ->template('12345', ['name' => 'Fabien']);
+
+        $message = $transport->send($mail);
+
+        $this->assertSame('foobar', $message->getMessageId());
+    }
+
+    public function testRemoteTemplateWithAlias()
+    {
+        $email = (new RemoteTemplateEmail())
+            ->template('welcome');
+        $envelope = new Envelope(new Address('fabpot@symfony.com', 'Fabien'), [new Address('saif.gmati@symfony.com', 'Saif Eddin')]);
+
+        $transport = new PostmarkApiTransport('KEY');
+        $method = new \ReflectionMethod(PostmarkApiTransport::class, 'getPayload');
+        $payload = $method->invoke($transport, $email, $envelope);
+
+        $this->assertSame('welcome', $payload['TemplateAlias']);
+        $this->assertArrayNotHasKey('TemplateId', $payload);
+        $this->assertEquals(new \stdClass(), $payload['TemplateModel']);
+    }
+
+    public function testRemoteTemplateRejectsSubject()
+    {
+        $email = (new RemoteTemplateEmail())
+            ->subject('Hello!')
+            ->template('welcome');
+        $envelope = new Envelope(new Address('fabpot@symfony.com', 'Fabien'), [new Address('saif.gmati@symfony.com', 'Saif Eddin')]);
+
+        $transport = new PostmarkApiTransport('KEY');
+        $method = new \ReflectionMethod(PostmarkApiTransport::class, 'getPayload');
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Postmark does not support overriding the subject of a template; define the subject in the template itself.');
+
+        $method->invoke($transport, $email, $envelope);
     }
 
     public function testSendThrowsForErrorResponse()