+++ src/Symfony/Component/Mailer/Bridge/Mailchimp/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add support for `RemoteTemplateEmail` to `MandrillApiTransport`
  * Add support for specifying the return path domain using the `X-MC-ReturnPathDomain` header
 
 7.4